Search by functional description statement values
This combines the search as it was before with an additional search by
functional description statement values. Due to the way statement values
are indexed by WikibaseCirrusSearch (as `"P123=SOMEVALUE"` strings) this
is done with a wildcard search, e.g. `P109=*MySearchTermGoesHere*`.

This approach has a few limitations:

* if the search term consists of multiple words, it is treated as a unit
  and not as individual terms for the functional description, i.e.
  search for "banana bread" will find "I like banana bread and kittens"
  but not "banana and bread"
* wildcard queries are considered expensive by Elasticsearch

Ideally this functional description value should be indexed separately
in Elasticsearch so that it can be queried more efficiently as mentioned
in the description of #79.

Fixes #79