feat: Make organization filter a tag selection
Allows for picking and also deleting an organization
using keyboard.
Includes a hack that always enforces a max. single
organization to be picked. Also makes the value
passed to the select field de facto incorrect --
as it is a multiselect nominally, the value
should be an array.
Working this around makes the filters to continue
work in search however.